Created reusable components for plots
