+++ client/src/pages/dashboards/KMeans.jsx
@@ -1,46 +1,14 @@
-import React, { useState, useEffect } from "react";
-import { DataGrid, GridToolbar } from '@mui/x-data-grid';
+import React from "react";
 import { Accordion, AccordionSummary, AccordionDetails, Typography } from '@mui/material';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import Plot from "react-plotly.js";
 import { useResizeDetector } from 'react-resize-detector';
+import ResponsivePlot from "../../components/plots/ResponsivePlot";
+import DataTable from "../../components/plots/DataTable";
+import ScatterPlot from "../../components/plots/ScatterPlot";
 
 function KMeans({ kmeansData }) {
-    const dataGridStyle = {
-        verflowY: 'auto',
-        '& .MuiDataGrid-columnHeaderTitle': {
-            fontWeight: 'bold',
-            fontSize: '17px',
-        },
-        '& .MuiDataGrid-row:nth-of-type(2n)': {
-            backgroundColor: '#f6f6f6',
-        },
-        '& .MuiDataGrid-toolbar': {
-            color: 'white',
-        },
-        '& .MuiButton-textPrimary': {
-            color: 'white !important',
-        },
-        '& .MuiTypography-root': {
-            color: 'white !important',
-        },
-        '& .MuiButtonBase-root': {
-            color: 'white !important',
-        },
-        '& .MuiSvgIcon-root': {
-            color: '#3fbdbd !important',
-        },
-        '& .MuiDataGrid-columnsManagement': {
-            backgroundColor: '#3fbdbd !important',
-        },
-    };
-
     const renderClusteredDataframe = () => {
         const keys = Object.keys(kmeansData.clustered_dataframe[0]);
-
-        // const orderedKeys = keys.includes('id') 
-        //   ? ['id', ...keys.filter((key) => key !== 'id')] 
-        //   : keys;
 
         const orderedKeys = keys.includes('id') 
             ? ['id', ...keys.filter((key) => key !== 'id' && key !== 'cluster'), 'cluster'] 
@@ -52,53 +20,13 @@ function KMeans({ kmeansData }) {
             flex: 1,
         }));
 
+        const rows = kmeansData.clustered_dataframe;
+
         return (
-            <DataGrid
-                rows={kmeansData.clustered_dataframe}
-                columns={cols}
-                loading={!kmeansData.clustered_dataframe.length}
-                showCellVerticalBorder
-                showColumnVerticalBorder
-                checkboxSelection={false}
-                initialState={{
-                    pagination: { paginationModel: { pageSize: 10 } },
-                }}
-                pageSizeOptions={[10, 25, 50]}
-                sx={dataGridStyle}
-                slots={{
-                    toolbar: GridToolbar,
-                }}
-                slotProps={{
-                    toolbar: {
-                        sx: {
-                            backgroundColor: '#474747',
-                            fontWeight: 'bold',
-                            padding: '10px',
-                            fontSize: '30px',
-                            color: '#ffffff',
-                            '& .MuiButtonBase-root': {
-                                color: 'white',
-                            },
-                        },
-                    },
-                }}
+            <DataTable
+                rows={rows}
+                cols={cols}
             />
-        );
-    };
-
-    const ResponsivePlot = ({ data, layout, config }) => {
-        const { width, height, ref } = useResizeDetector();
-
-        return (
-            <div ref={ref} style={{ width: '100%', height: '500px' }}>
-                {width && height && (
-                    <Plot
-                        data={data}
-                        layout={{ ...layout, width, height }}
-                        config={config}
-                    />
-                )}
-            </div>
         );
     };
 
@@ -132,42 +60,11 @@ function KMeans({ kmeansData }) {
         }));
 
         return (
-            <ResponsivePlot
+            <ScatterPlot
                 data={data}
-                layout={{
-                    autosize: true,
-                    title: {
-                        text: 'Clusters scatter plot',
-                    },
-                    xaxis: {
-                        title: {
-                            text: 'x',
-                        },
-                        automargin: true,
-                        showgrid: true,
-                        zeroline: false,
-                    },
-                    yaxis: {
-                        title: {
-                            text: 'y',
-                        },
-                        automargin: true,
-                        showgrid: true,
-                        zeroline: false,
-                    },
-                    legend: {
-                        orientation: 'h',
-                        x: 0.5,
-                        xanchor: 'center',
-                        y: -0.2,
-                    },
-                    hovermode: 'closest',
-                }}
-                config={{
-                    responsive: true,
-                    displayModeBar: true,
-                    displaylogo: false, 
-                }}
+                title={'Clusters scatter plot'}
+                xTitle={'x'}
+                yTitle={'y'}
             />
         );
     };
@@ -201,45 +98,12 @@ function KMeans({ kmeansData }) {
             },
         }));
 
-        console.log(data)
-
         return (
-            <ResponsivePlot
+            <ScatterPlot
                 data={data}
-                layout={{
-                    autosize: true,
-                    title: {
-                        text: 'Clusters scatter plot',
-                    },
-                    xaxis: {
-                        title: {
-                            text: 'x',
-                        },
-                        automargin: true,
-                        showgrid: true,
-                        zeroline: false,
-                    },
-                    yaxis: {
-                        title: {
-                            text: 'y',
-                        },
-                        automargin: true,
-                        showgrid: true,
-                        zeroline: false,
-                    },
-                    legend: {
-                        orientation: 'h',
-                        x: 0.5,
-                        xanchor: 'center',
-                        y: -0.2,
-                    },
-                    hovermode: 'closest',
-                }}
-                config={{
-                    responsive: true,
-                    displayModeBar: true,
-                    displaylogo: false, 
-                }}
+                title={'Scatter plot for classes'}
+                xTitle={'x'}
+                yTitle={'y'}
             />
         );
     };
@@ -271,62 +135,23 @@ function KMeans({ kmeansData }) {
             flex: 1,
         }));
 
+        const rows = kmeansData.centroids;
+
         return (
-            <DataGrid
-                rows={kmeansData.centroids}
-                columns={cols}
-                loading={!kmeansData.centroids.length}
-                sx={dataGridStyle}
+            <DataTable
+                rows={rows}
+                cols={cols}
             />
         );
     };
 
     return (
         <div>
-            <div style={{ 
-                display: 'flex',
-                justifyContent: 'center',
-                width: '100%',
-                marginBottom: '30px'
-            }}>
-                {renderClusteredDataframe()}
-            </div>
-
-            <div style={{ 
-                display: 'flex',
-                justifyContent: 'center',
-                width: '100%',
-                marginBottom: '30px'
-            }}>
-                {renderScatterPlot()}
-            </div>
-
-            <div style={{ 
-                display: 'flex',
-                justifyContent: 'center',
-                width: '100%',
-                marginBottom: '30px'
-            }}>
-                {renderScatterPlotForSpecies()}
-            </div>
-
-            <div style={{ 
-                display: 'flex',
-                justifyContent: 'center',
-                width: '100%',
-                marginBottom: '30px'
-            }}>
-                {renderBarPlot()}
-            </div>
-
-            <div style={{ 
-                display: 'flex',
-                justifyContent: 'center',
-                width: '100%',
-                marginBottom: '30px'
-            }}>
-                {renderCentroids()}
-            </div>
+            {renderClusteredDataframe()}
+            {renderScatterPlot()}
+            {renderScatterPlotForSpecies()}
+            {renderBarPlot()}
+            {renderCentroids()}
         </div>
     )
 }
